update app.js
In your older version, you can able to add empty Dos but this is not a good practice. So I add one simple condition to check whether 'name' is empty or not if the name is not empty then simply add it to the array but if 'name' is empty then we do not add to the array.